# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ynovee Hotel Booking API",
        "description": "REST API for the Ynovee Hotel Booking System. Supports room management, bookings, Stripe payments, and content management.",
        "contact": {
            "email": "noreply@ynovee.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "List all room types",
                "description": "Returns all room types. Optionally filter by availability and guest capacity.",
                "operationId": "bea681612754126f01807ff95647eb34",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Check-in date for availability filter",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Check-out date for availability filter",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "guests",
                        "in": "query",
                        "description": "Minimum guest capacity",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of room types",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RoomType"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Create a room type (Admin)",
                "operationId": "cd6c1029f22e2425a1f99cf83e27383a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "capacity",
                                    "price_usd",
                                    "price_eur",
                                    "price_ghs",
                                    "total_rooms"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Deluxe Suite"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "price_usd": {
                                        "type": "number",
                                        "example": 150
                                    },
                                    "price_eur": {
                                        "type": "number",
                                        "example": 140
                                    },
                                    "price_ghs": {
                                        "type": "number",
                                        "example": 2500
                                    },
                                    "total_rooms": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "amenities": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Room type created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomType"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/rooms/{id}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get a room type",
                "operationId": "91a59d880bb0297e8321dd75af8f506b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room type details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomType"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Update a room type (Admin)",
                "operationId": "c489811fdd25dfe273cf9a5bc3b5d386",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "capacity",
                                    "price_usd",
                                    "price_eur",
                                    "price_ghs",
                                    "total_rooms"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "price_usd": {
                                        "type": "number"
                                    },
                                    "price_eur": {
                                        "type": "number"
                                    },
                                    "price_ghs": {
                                        "type": "number"
                                    },
                                    "total_rooms": {
                                        "type": "integer"
                                    },
                                    "amenities": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room type updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomType"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Delete a room type (Admin)",
                "operationId": "61024891be80949c0aaa6e4536381225",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "RoomType": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "example": "Deluxe Ocean View"
                    },
                    "description": {
                        "type": "string"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 2
                    },
                    "price_usd": {
                        "type": "number",
                        "format": "float",
                        "example": 150
                    },
                    "price_eur": {
                        "type": "number",
                        "format": "float",
                        "example": 140
                    },
                    "price_ghs": {
                        "type": "number",
                        "format": "float",
                        "example": 2500
                    },
                    "total_rooms": {
                        "type": "integer",
                        "example": 5
                    },
                    "amenities": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the Bearer token obtained from POST /api/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}